♻️ [setting/location/map] rewrite to composition API
